refactor(unpack): 目录包化并收口命名
Why:
- 当前 unpack.py 体积过大，职责边界不清晰
- 需要为后续全项目统一命名收口保留兼容过渡层

What:
- 将 unpack.py 拆分为 unpack 目录包，并按 batch、entity、wav、bp_vo 划分职责
- 将仓内主调用切换到 unpack_all、unpack_entity、unpack_map、execute_tasks 等新命名，并保留旧名兼容别名
- 更新 facade 与相关测试，补充 unpack 公开导出面回归测试

Risk:
- 兼容性影响：有，公开 API 主名已切换，但旧名别名仍可继续使用
- 潜在风险：后续统一移除兼容层时仍需同步外部调用方

Test:
- uv run ruff check src/lol_audio_unpack/unpack src/lol_audio_unpack/facade.py tests/test_unpack_package_api.py tests/test_unpack_bp_vo.py tests/test_context_injection.py tests/test_unpack_wav_sidecar.py tests/test_remote_preparer.py
- uv run pytest -q tests/test_unpack_package_api.py tests/test_unpack_bp_vo.py tests/test_context_injection.py tests/test_unpack_wav_sidecar.py tests/test_wav_sidecar.py tests/test_remote_preparer.py::test_facade_extract_prepares_remote_wads_before_unpack
- uv run python -c "from lol_audio_unpack import unpack; print(callable(unpack.unpack_all), callable(unpack.unpack_audio_all), unpack.unpack_all is unpack.unpack_audio_all)"
